feature: ADD focuspoint/openai get generate content
